Allow loading of redis dumps with binary data.
If the redis dump contains binary data (e.g. serialized data) that
appears to be invalid UTF8 to the YAJL parser, it will not load.
(https://github.com/delano/redis-dump/issues/7). I've added a "-n"
flag that will tell YAJL to ignore "bad utf8" in the case where the
user knows that their data is binary in nature.